% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd_3}
\alias{cd_3}
\title{Longitudinal CD4 Lymphocyte Counts for HIV Patients (2018-2024)}
\format{
A data frame with 176 rows and 18 variables:
\describe{
\item{ID}{Unique identifier for each patient.}
\item{cd_2018_1 to cd_2024_3}{Longitudinal CD4 lymphocyte count
measurements. The naming convention follows the year (from 2018 to 2024)
and the measurement period within the year (1, 2, or 3). Values are
recorded in cells per micro liter of blood.}
}
}
\source{
Clinical data from Hospital Vicente Guerrero, IMSS, HIV Clinic.
}
\usage{
cd_3
}
\description{
Contains longitudinal measurements of CD4 lymphocyte counts for
176 patients living with HIV, recorded over the period from 2018 to 2024.
CD4 counts are a critical indicator of immune function, used to monitor the
progression of HIV and the effectiveness of treatments. Measurements were
taken at various points throughout the study, with some missing values due
to unavailable data for specific patients at certain times.
}
\details{
.
CD4 counts are used to monitor immune system health in individuals with
HIV. A lower CD4 count often indicates a weakened immune system, whereas
higher counts suggest a stronger immune response. Some values are missing,
indicating no measurement was taken for a particular patient at that time.
}
\examples{
# Load the dataset
data(cd_3)

# Summarize CD4 counts for the year 2021
summary(cd_3[, c("cd_2021_1", "cd_2021_2", "cd_2021_3")])

}
\keyword{datasets}
