% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_check.R
\name{complex_check}
\alias{complex_check}
\title{Check if Coefficients of a Qubit State Object are Complex Numbers}
\usage{
complex_check(qstate_obj)
}
\arguments{
\item{qstate_obj}{A qubit state object created using the qsimulatR package.}
}
\value{
Coefficients class of given qubit state object
}
\description{
This function returns the class of the coefficients of a given qubit state object
}
\examples{
library(qsimulatR)
ket1 <- qstate(nbits = 1, coefs = c(0, 1))
complex_check(ket1)
}
