% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactions.R
\name{create_interactions}
\alias{create_interactions}
\title{Create Interactions between longitudinal values of CD4 Lymphocytes and Viral
Load Data}
\usage{
create_interactions(cd_data, vl_data)
}
\arguments{
\item{cd_data}{A data frame containing longitudinal CD4 count values for patients.}

\item{vl_data}{A data frame containing longitudinal viral load values for patients.}
}
\value{
A list with the following components:
\describe{
\item{cd_diff}{A tibble containing the differences in CD4 count values for each patient.}
\item{cds_diff}{A tibble containing the standardized differences in CD4 count values.}
\item{vl_diff}{A tibble containing the differences in viral load values for each patient.}
\item{vlog_diff}{A tibble containing the log10-transformed differences in viral load values.}
\item{vlogs_diff}{A tibble containing the standardized log10-transformed differences in viral load values.}
}
}
\description{
This function computes the differences in CD4 counts and viral loads over time
for a set of patients, standardizes the data, and returns a structured output
containing the calculated differences.
}
\examples{
data(cd_3)
data(vl_3)
create_interactions(cd_3, vl_3)
}
