% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.payoffs.R
\name{mse.payoffs}
\alias{mse.payoffs}
\title{Mean Squared Errors for Payoff Predictions}
\usage{
\method{mse}{payoffs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{payoffs}.}

\item{...}{Additional arguments passed to other methods (currently not used).}
}
\description{
Computes the mean squared error (MSE) between observed CD4 and viral load differences
and their corresponding predicted payoff values within each interaction classification.
}
