% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.InteractionClassification.R
\name{plot.InteractionClassification}
\alias{plot.InteractionClassification}
\title{Plot InteractionClassification Clusters}
\usage{
\method{plot}{InteractionClassification}(x, ...)
}
\arguments{
\item{x}{An object of class \code{InteractionClassification} created using the
\code{InteractionClassification()} function. It contains the clustering results and cluster centers.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A ggplot object that visualizes the clusters and their centers.
}
\description{
This function visualizes the clusters formed by the \code{InteractionClassification} object.
It displays the viral load differences (\code{vl_diff}) and CD4 count differences (\code{cd_diff})
as points, with each point color-coded by its cluster. Cluster centers are shown as red stars.
}
\examples{
data(vl_3)
data(cd_3)
interaction_obj <- create_interactions(cd_3[,-1], vl_3[,-1])
class_obj <- InteractionClassification(interaction_obj$vlogs_diff, interaction_obj$cds_diff)
plot(class_obj)
}
