% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preds2}
\alias{preds2}
\title{Batched Predictions for Longitudinal Viral Load Values for HIV Patients (2018-2024)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 176 rows and 1 columns.
}
\source{
Clinical data from Hospital Vicente Guerrero, IMSS, HIV Clinic.
}
\usage{
preds2
}
\description{
Contains batched predictions of longitudinal viral load values for 176
patients from 2018 to 2024.
}
\examples{
data(preds2)
head(preds2)

}
\keyword{datasets}
