% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.interaction.R
\name{print.interaction}
\alias{print.interaction}
\title{Print Summary of CD4 and Viral Load Differences}
\usage{
\method{print}{interaction}(x, ...)
}
\arguments{
\item{x}{An object of class \code{interaction} containing CD4 and viral load
differences.}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
This function does not return a value; it prints the summary
statistics directly to the console.
}
\description{
This method prints a summary of the CD4 and viral load differences contained
in the interaction object. It provides both raw and standardized values for
CD4 differences, as well as raw, log10-transformed, and log10-standardized
values for viral load differences.
}
\examples{
data("vl_3")
data("cd_3")
interaction_obj <- create_interactions(cd_3, vl_3)
print(interaction_obj)
}
