% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.InteractionClassification.R
\name{print.InteractionClassification}
\alias{print.InteractionClassification}
\title{Print Method for InteractionClassification Objects}
\usage{
\method{print}{InteractionClassification}(x, ...)
}
\arguments{
\item{x}{An object of class \code{InteractionClassification} containing
the results of the k-means clustering.}

\item{...}{Additional arguments that may be passed to the print method.}
}
\value{
The function does not return a value; it prints the summary
information to the console.
}
\description{
This method prints a summary of the InteractionClassification object,
including the number of clusters, the cluster means, and the sizes of
each cluster.
}
\examples{
data(vl_3)
data(cd_3)
interaction_obj <- create_interactions(cd_3[,-1], vl_3[,-1])
class_obj <- InteractionClassification(interaction_obj$vlogs_diff, interaction_obj$cds_diff)

# Print the summary of the classification
print(class_obj)
}
