% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvirus-package.R
\docType{package}
\name{qvirus-package}
\alias{qvirus}
\alias{qvirus-package}
\title{qvirus: Quantum Computing for HIV Drug Design}
\description{
Resources, tutorials, and code snippets dedicated to exploring the intersection of quantum computing and artificial intelligence (AI) in the context of HIV drug design. With the emergence of quantum artificial intelligence (QAI) and the development of small-scale quantum computers, there's an unprecedented opportunity to revolutionize the way antiretroviral therapy design for HIV is approached. This repository serves as a platform to discuss, develop, and share insights into leveraging quantum computing techniques, particularly through the use of Qiskit and R programming, to address the challenges in designing effective inhibitors of HIV protease enzymes.
}
\author{
\strong{Maintainer}: Juan Pablo Acuña González \email{acua6307@gmail.com} (\href{https://orcid.org/0009-0003-6029-6560}{ORCID})

}
\keyword{internal}
