% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDrugProteinBinding.R
\name{simulateDrugProteinBinding}
\alias{simulateDrugProteinBinding}
\title{Simulate Drug-Protein Binding}
\usage{
simulateDrugProteinBinding(drug_state, protein_state, entanglement_type)
}
\arguments{
\item{drug_state}{Quantum state representing the drug molecule.}

\item{protein_state}{Quantum state representing the protein molecule.}

\item{entanglement_type}{Type of entanglement to use: "simple" for H + CNOT gates or "separable" for kronecker + CNOT + RX gates.}
}
\value{
A list containing:
\item{entangled_state}{The entangled quantum state representing drug-protein interaction.}
\item{measured_state}{The measured state after simulating the drug-protein binding.}
}
\description{
This function simulates the binding of a drug and a protein using quantum states and
different types of entanglement, illustrating their effect in the context of HIV drug design.
}
\examples{
library(qsimulatR)
# Define Drug and Protein states
Drug_State <- qstate(1, coefs = as.complex(c(1, 0)))  # State |0>
Protein_State <- qstate(1, coefs = as.complex(c(1, 0)))  # State |0>

# Simulate drug-protein binding with separable states (kronecker + CNOT + RX)
result_separable <- simulateDrugProteinBinding(Drug_State, Protein_State, 
entanglement_type = "separable")
result_separable$measured_state
result_separable$entangled_state
}
