% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.payoffs.R
\name{summary.payoffs}
\alias{summary.payoffs}
\title{Summarize Payoffs}
\usage{
\method{summary}{payoffs}(object, ...)
}
\arguments{
\item{object}{A payoffs object.}

\item{...}{Additional arguments (not used).}
}
\description{
This function summarizes the payoffs object by classification.
}
