% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vl_3}
\alias{vl_3}
\title{Longitudinal Viral Load Values for HIV Patients (2018-2024)}
\format{
A data frame with 176 rows and 18 variables:
\describe{
\item{ID}{Unique identifier for each patient.}
\item{vl_2018_1 to vl_2024_3}{Viral load measurements taken at various
points in time between 2018 and 2024. The naming structure represents the
year of the measurement (e.g., 2018, 2019, etc.), and the point in time
within the year (1, 2, or 3). Units are in copies of HIV RNA per milliliter
of blood.}
}
}
\source{
Clinical data from Hospital Vicente Guerrero, IMSS, HIV Clinic.
}
\usage{
vl_3
}
\description{
Contains longitudinal measurements of viral load for 176
patients from 2018 to 2024. Viral load is a critical marker used to monitor
the effectiveness of HIV treatment by measuring the amount of HIV RNA in the
blood.
}
\details{
The viral load measurements provide insight into the patient's response to
antiretroviral therapy (ART). Lower viral load values, especially
undetectable levels, indicate better control of the infection. Missing
values indicate that no viral load measurement was available for that patient
at that specific time.
}
\examples{
\dontrun{
# Load the dataset
data(vl_3)

# Summarize viral loads for the year 2021
summary(vl_3[, c("cd_2021_1", "cd_2021_2", "cd_2021_3")])
}

}
\keyword{datasets}
