% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.InteractionClassification.R
\name{mse.InteractionClassification}
\alias{mse.InteractionClassification}
\title{Mean Squared Errors method for the InteractionClassification class}
\usage{
\method{mse}{InteractionClassification}(object, ...)
}
\arguments{
\item{object}{An object of class \code{InteractionClassification}.}

\item{...}{Additional arguments (currently not used).}
}
\description{
This method computes the mean squared errors (MSE) for viral load and CD4 differences
based on the classification results from an InteractionClassification object.
}
