% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.InteractionClassification.R
\name{summary.InteractionClassification}
\alias{summary.InteractionClassification}
\title{Summarize an InteractionClassification object}
\usage{
\method{summary}{InteractionClassification}(object, ...)
}
\arguments{
\item{object}{An object of class \code{InteractionClassification} returned by the
\code{\link[=InteractionClassification]{InteractionClassification()}} function.}

\item{...}{Additional arguments passed to other methods (currently not used).}
}
\description{
Computes summary statistics by classification group from an object of class
\code{InteractionClassification}, including mean differences in viral load and CD4 counts,
and the number of observations per cluster.
}
